Move Join Request to Entities Folder
