+++ src/main/java/com/kevin/prayerappservice/join/JoinRequestRepository.java
@@ -1,5 +1,6 @@
 package com.kevin.prayerappservice.join;
 
+import com.kevin.prayerappservice.join.entities.JoinRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
